fix(server): 修复代码审查报告 P1/P2/P3 共 34 项 BUG
- P1(4): 运行时配置覆盖失效(改 LiveMemoryConfigurationSource)、匹配竞态重复建房、房间密码被清空、Bots 假响应
- P2(18): 密码迭代上限、防重放清理、Token uid 转义、好友/拉黑唯一键规范化、帧长/消息体上限、TakeDamage 溢出、Logger HMAC 鉴权+限流、HTTP 管理接口 Token 绑定与假成功整改、DB 空表守卫等
- P3(12): 帧解析/属性编解码限长/KCP per-IP O(1)/路由字段溢出/Supervisor 空日志/脚本 tick 串行化/AOI 非玩家投递/World 房间复用/OrderedTaskQueue ABA 等
- 验证: 全部 26 个项目编译通过(0 警告 0 错误)

+++ NetGameServer/Framework/Framework.Scripting/IEntityScript.cs
@@ -145,11 +145,12 @@ public abstract class EntityScriptBase : IEntityScript
     /// </summary>
     public static int MathClampAdd(EntityObj entity, string name, int delta, int min, int max)
     {
-        int newValue = entity.Get<int>(name) + delta;
+        // P2 修复：内部用 long 累加，防止 Get<int> + delta 整数溢出回绕（恶意大 delta 可导致结果跳变）。
+        long newValue = (long)entity.Get<int>(name) + delta;
         if (newValue < min) newValue = min;
         else if (newValue > max) newValue = max;
-        entity.Set(name, newValue);
-        return newValue;
+        entity.Set(name, (int)newValue);
+        return (int)newValue;
     }
 }
